- Factored out the (not commonly used) feature of tracking which individual worker is from which factory.  Should make the main body of the manager easier to understand and modify.
